+++ engine/src/prelude.rs
@@ -4,17 +4,27 @@
 //! ```
 //! - imports everything this engine has to offer.
 
-pub use crate::{camera::*, directions::*, objects::*, resources::*, *};
+#[cfg(feature = "client")]
+pub use crate::resources::*;
+pub use crate::{camera::*, directions::*, objects::*, *};
+#[cfg(feature = "client")]
 pub use appearance::*;
+#[cfg(feature = "client")]
 pub use color::*;
+#[cfg(feature = "client")]
 pub use data::*;
+#[cfg(feature = "client")]
 pub use events::*;
 pub use glam;
 pub use glam::{vec2, Vec2};
 pub use joints::*;
+#[cfg(feature = "client")]
 pub use labels::*;
+#[cfg(feature = "client")]
 pub use materials::*;
 pub use physics::*;
 pub use scenes::*;
+#[cfg(feature = "client")]
 pub use textures::*;
+#[cfg(feature = "client")]
 pub use window::*;